+++ src/phase1/branch_analysis.rs
@@ -6,21 +6,21 @@ use yaxpeax_arm::armv7::Reg;
 #[derive(Debug)]
 pub enum JumpType {
     /// BL/BLX with immediate
-    DirectCall(usize),
+    DirectCall(u32),
     /// BLX with register
     IndirectCall(Reg),
     /// unconditional B
-    DirectJump(usize),
+    DirectJump(u32),
     /// BX
     IndirectJump(Reg),
     /// conditional B/CBZ/CBNZ
-    Branch { target: usize, fallthrough: usize },
+    Branch { target: u32, fallthrough: u32 },
 }
 
 #[derive(Debug)]
 pub struct BranchAnalysis {
     /// caller -> callee (even though it's not always call, but we're not going to use something like jumper and jumpee, right?)
-    pub jumps: HashMap<usize, JumpType>,
+    pub jumps: HashMap<u32, JumpType>,
 }
 
 impl BranchAnalysis {
@@ -32,31 +32,31 @@ impl BranchAnalysis {
 
     /// mark `target` as function call with `me` as caller VA
     #[instrument(skip(self), fields(me = format_args!("{:#x}", me), target = format_args!("{:#x}", target)), level = "trace")]
-    pub fn mark_as_direct_call(&mut self, me: usize, target: usize) {
+    pub fn mark_as_direct_call(&mut self, me: u32, target: u32) {
         self.jumps.insert(me, JumpType::DirectCall(target));
     }
 
     /// mark `target` as function call with `me` as caller VA
     #[instrument(skip(self), fields(me = format_args!("{:#x}", me)), level = "trace")]
-    pub fn mark_as_indirect_call(&mut self, me: usize, reg: Reg) {
+    pub fn mark_as_indirect_call(&mut self, me: u32, reg: Reg) {
         self.jumps.insert(me, JumpType::IndirectCall(reg));
     }
 
     /// mark `target` as direct jump with `me` as jump instruction VA
     #[instrument(skip(self), fields(me = format_args!("{:#x}", me), target = format_args!("{:#x}", target)), level = "trace")]
-    pub fn mark_as_direct_jump(&mut self, me: usize, target: usize) {
+    pub fn mark_as_direct_jump(&mut self, me: u32, target: u32) {
         self.jumps.insert(me, JumpType::DirectJump(target));
     }
 
     /// mark `target` as indirect jump with `me` as jump instruction VA
     #[instrument(skip(self), fields(me = format_args!("{:#x}", me)), level = "trace")]
-    pub fn mark_as_indirect_jump(&mut self, me: usize, reg: Reg) {
+    pub fn mark_as_indirect_jump(&mut self, me: u32, reg: Reg) {
         self.jumps.insert(me, JumpType::IndirectJump(reg));
     }
 
     /// mark `target` as branch target arm with `me` as branch instruction VA
     #[instrument(skip(self), fields(me = format_args!("{:#x}", me), target = format_args!("{:#x}", target), fallthrough = format_args!("{:#x}", fallthrough)), level = "trace")]
-    pub fn mark_as_branch(&mut self, me: usize, target: usize, fallthrough: usize) {
+    pub fn mark_as_branch(&mut self, me: u32, target: u32, fallthrough: u32) {
         self.jumps.insert(
             me,
             JumpType::Branch {
@@ -70,20 +70,20 @@ impl BranchAnalysis {
     ///
     /// if it's a function, then at least one `JumpType::DirectCall` should point to it
     #[instrument(skip(self), fields(va = format_args!("{:#x}", va)), level = "trace")]
-    pub fn is_fn(&self, va: usize) -> bool {
+    pub fn is_fn(&self, va: u32) -> bool {
         self.jumps.values().any(|ty| match ty {
             JumpType::DirectCall(v) => va == *v,
             _ => false,
         })
     }
 
-    pub fn get_callee(&self, va: usize) -> Option<&JumpType> {
+    pub fn get_callee(&self, va: u32) -> Option<&JumpType> {
         self.jumps.get(&va)
     }
 
     /// get all jumps to the `va`, either it's a branch or jump
     #[instrument(skip(self), fields(va = format_args!("{:#x}", va)), level = "trace")]
-    pub fn all_jumps_for(&self, va: usize) -> impl Iterator<Item = usize> {
+    pub fn all_jumps_for(&self, va: u32) -> impl Iterator<Item = u32> {
         self.jumps
             .iter()
             .filter_map(move |(caller_va, ty)| match ty {
@@ -103,7 +103,7 @@ impl BranchAnalysis {
 
     /// get all jumps, branches and calls for the `va`
     #[instrument(skip(self), fields(va = format_args!("{:#x}", va)), level = "trace")]
-    pub fn all_for(&self, va: usize) -> impl Iterator<Item = usize> {
+    pub fn all_for(&self, va: u32) -> impl Iterator<Item = u32> {
         self.jumps
             .iter()
             .filter_map(move |(caller_va, ty)| match ty {
@@ -125,7 +125,7 @@ impl BranchAnalysis {
 
     /// remove entry
     #[instrument(skip(self), fields(va = format_args!("{:#x}", va)), level = "trace")]
-    pub fn discard(&mut self, va: usize) {
+    pub fn discard(&mut self, va: u32) {
         self.jumps.remove_entry(&va);
     }
 }
